ensure there are no issues with concatenation;  [touch:7792]

+++ tests/testcases/core/helpers/EEH_DTT_Helper_Test.php
@@ -242,7 +242,7 @@ class EEH_DTT_Helper_Test extends EE_UnitTestCase {
 		//now in timezone currently set.
 		$default_timezone = new DateTimeZone( EEH_DTT_Helper::get_timezone() );
 		$now = new DateTime( "now",  $default_timezone );
-		$expected_offset = $now->format( 'U' ) + timezone_offset_get( $default_timezone, $now );
+		$expected_offset = (int)$now->format( 'U' ) + (int)timezone_offset_get( $default_timezone, $now );
 
 		$this->assertEquals( $expected_offset, EEH_DTT_Helper::get_timestamp_with_offset( $now->format('U' ) ) );
 
@@ -251,7 +251,7 @@ class EEH_DTT_Helper_Test extends EE_UnitTestCase {
 
 		//now let's test with a different timezone for the incoming timestamp.
 		$now->setTimeZone( new DateTimeZone( 'America/Toronto' ) );
-		$expected_timestamp = $now->format('U') + timezone_offset_get( new DateTimeZone( 'America/Toronto' ), $now );
+		$expected_timestamp = (int)$now->format('U') + (int)timezone_offset_get( new DateTimeZone( 'America/Toronto' ), $now );
 		$this->assertEquals( $expected_timestamp, EEH_DTT_Helper::get_timestamp_with_offset( $now->format('U'), 'America/Toronto' ) );
 	}
